AhmadHafizuddinRahmat_A23MJ5070/backend/ocr_service.py
import pytesseract
from PIL import Image
import io

# Default Windows install path — change if yours is different
pytesseract.pytesseract.tesseract_cmd = r"C:\Program Files\Tesseract-OCR\tesseract.exe"

def extract_text_from_image(image_bytes: bytes) -> str:
    image = Image.open(io.BytesIO(image_bytes))
    if image.mode not in ("RGB", "L"):
        image = image.convert("RGB")
    return pytesseract.image_to_string(image, lang="eng").strip()