AhmadHafizuddinRahmat_A23MJ5070/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
sqlalchemy==2.0.35
python-docx==1.1.2
extract-msg==0.48.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1